refactor job lisitng, progress on detail
