+++ src/pages/Home.jsx
@@ -1,3 +1,9 @@
+import JobCard from "../components/JobCard";
+
 export default function Home() {
-  return <div>Homepage</div>;
+  return (
+    <div>
+      <JobCard />
+    </div>
+  );
 }
